File the route guard where a source-text scan belongs, and update the register
Follow-up to the KBR-127 fix, from a second design review of the specification.
Three findings, one of which was an error in my own reasoning.

**The structural guard moves to L2, in its own file.** It was written inside the
behavioural file, taking the `l1` path default, on a reading of TEST_SUITE.md §6.2's
scope line — "two artifacts that must agree but are deployed, edited or upgraded
separately" — that a scan of one source tree is not two artifacts. The repo's own
operative reading says otherwise, in the file §6.2 names as the pattern to copy:
`tests/test_egress_coverage.py` classifies itself as "an artifact outside `src/kitty`
Python code, **or a structural scan of source text**", and gives a second reason the
scope argument never addressed — it "keeps a source-text scan out of the L1 set that
mutation testing will judge". §6.1 makes mutation testing L1's validation, so a
source-text scan there has its kills attributed to the wrong layer. The sibling guard
this ticket pairs with, `test_wire_shape_honesty.py` (KBR-7), is L2 for the same reason.

`test_layer_selection.py` pins the L2 file list in both directions and went red until
the new file was registered. That is the check working as designed, so the entry is
added rather than the marker dropped.

**The guard now forbids `_backends` as well.** `_backends[idx][2].model` reaches the
same profile string without naming `_active_model` or `_model`, so a helper could have
regained the second source of truth and passed both directions of the guard. Positive
control added for that shape.

**`translate_to_upstream` must not remove `model` from the request in place.** Fourteen
of the twenty-three `_build_upstream_url` sites run *after* serialization on the same
dict, on failover paths, so KBR-127 turned that from an implementation detail into an
invariant. `AzureOpenAIAdapter` is the one adapter that removes `model` (register entry
P6) and is safe only because it builds a copy. Pinned across all 23 registered adapters.

**Two claims in the task requirements were wrong, and the docstring that repeated one
is corrected.** I had written that every `_build_upstream_url` call site is preceded
within its own function by `_normalize_model`, with `_make_upstream_request` as the lone
exception. That came from comparing line numbers. Under control flow it is false: all
five methods take `cc_request` as a parameter, none normalizes unconditionally in its
own body, and every `_normalize_model` that lexically precedes the first helper call
sits inside the `if use_custom_transport:` block that returns or falls through. The
invariant that actually holds is one level up — each of the five is handed an
already-normalized request by its caller. Separately, `_request_with_retry_balancing`
does re-normalize after every `_select_backend()`; only `_request_with_retry_single`
relies on its caller. `_make_upstream_request`'s docstring now states the precondition
and which wrapper satisfies it how.

Never settle a control-flow question with a line-order comparison.

**TEST_SUITE.md's register and routing oracle are updated.** Entry M14 ("the destination
is built from the profile"), entry P20 ("the profile's model as the deployment id"), and
§3.3.5's Azure row and its "derive the routing expectation from the configured profile"
rule all describe a source this change replaces: the destination now resolves from the
request's normalized model — `normalize_model_name(profile.model)` where a profile names
one, the agent's model where it does not. §3.2.2 makes the register the specification, so
leaving the rows would hand a wrong derivation rule to whoever implements the §3.3.5
oracle. A seventh falsification control is added for the prefixed-profile case, because
path, auth and body each pass when checked alone.
